+++ example/android/javalib/4-sum-lib-java/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.calculator.app" android:versionCode="1" android:versionName="1.0">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.example.app" android:versionCode="1" android:versionName="1.0">
     <uses-sdk android:minSdkVersion="9" android:targetSdkVersion="35"/>
     <application android:label="Calculator" android:theme="@android:style/Theme.Light.NoTitleBar" android:debuggable="true">
         <activity android:name=".Main"
