+++ example/android/javalib/4-sum-lib-java/app/src/test/java/com/example/CalculatorUnitTest.java
@@ -1,4 +1,4 @@
-package com.calculator.app;
+package com.example.app;
 
 import static org.junit.Assert.*;
