+++ example/android/javalib/4-sum-lib-java/build.mill
@@ -40,7 +40,6 @@ object lib extends AndroidLibModule with PublishModule {
   )
 
   // Configuration for ReleaseKey
-  def androidReleaseKeyName: T[Option[String]] = Task { Some("releaseKey.jks") }
   def androidReleaseKeyAlias: T[Option[String]] = Task { Some("releaseKey") }
   def androidReleaseKeyPass: T[Option[String]] = Task { Some("MillBuildTool") }
   def androidReleaseKeyStorePass: T[Option[String]] = Task { Some("MillBuildTool") }
