+++ example/android/javalib/4-sum-lib-java/lib/src/main/java/com/example/Sum.java
@@ -1,4 +1,4 @@
-package com.sumlib;
+package com.example;
 
 public class Sum {
   public static int apply(int[] numbers) {
