+++ example/android/javalib/4-sum-lib-java/lib/src/test/java/com/example/SumUnitTest.java
@@ -1,4 +1,4 @@
-package com.sumlib;
+package com.example;
 
 import static org.junit.Assert.*;
