+++ example/android/kotlinlib/4-sum-lib-kotlin/app/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.helloworld.app" android:versionCode="1" android:versionName="1.0">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.example.app" android:versionCode="1" android:versionName="1.0">
     <uses-sdk android:minSdkVersion="9" android:targetSdkVersion="35"/>
     <application android:label="@string/app_name" android:theme="@android:style/Theme.Light.NoTitleBar" android:debuggable="true">
         <activity android:name=".MainActivity"
@@ -9,8 +9,4 @@
             </intent-filter>
         </activity>
     </application>
-    <!-- TODO this needs to be constructed by Mill's integration tests (android tests) -->
-    <instrumentation
-            android:name="androidx.test.runner.AndroidJUnitRunner"
-            android:targetPackage="com.helloworld.app" />
 </manifest>
