+++ example/android/kotlinlib/4-sum-lib-kotlin/app/src/main/java/com/example/SampleLogic.kt
@@ -1,4 +1,4 @@
-package com.calculator
+package com.example
 
 object SampleLogic {
     fun textSize(): Float = 32f
