+++ example/android/kotlinlib/4-sum-lib-kotlin/app/src/main/kotlin/com/example/SampleLogicInKotlinDir.kt
@@ -1,6 +1,6 @@
-package com.calculator
+package com.example
 
-import com.sumlib.Sum
+import com.example.Sum
 
 object SampleLogicInKotlinDir {
