+++ example/android/kotlinlib/4-sum-lib-kotlin/app/src/test/kotlin/com/example/ExampleUnitTestInKotlinDir.kt
@@ -1,4 +1,4 @@
-package com.calculator
+package com.example
 
 import org.junit.Assert.*
 import org.junit.Test
