+++ example/android/kotlinlib/4-sum-lib-kotlin/lib/src/main/kotlin/com/example/Sum.kt
@@ -1,4 +1,4 @@
-package com.sumlib
+package com.example
 
 object Sum {
     fun apply(numbers: Array<Int>): Int {
