+++ kotlinlib/src/mill/kotlinlib/android/AndroidAppKotlinModule.scala
@@ -27,37 +27,12 @@ import mill.scalalib.TestModule.Junit5
  * [[https://developer.android.com/studio Android Studio Documentation]]
  */
 @mill.api.experimental
-trait AndroidAppKotlinModule extends AndroidAppModule with KotlinModule { outer =>
+trait AndroidAppKotlinModule extends AndroidAppModule with AndroidKotlinModule { outer =>
 
   override def sources: T[Seq[PathRef]] =
     super[AndroidAppModule].sources() :+ PathRef(moduleDir / "src/main/kotlin")
 
-  override def kotlincOptions = super.kotlincOptions() ++ {
-    if (androidEnableCompose()) {
-      Seq(
-        // TODO expose Compose configuration options
-        // https://kotlinlang.org/docs/compose-compiler-options.html possible options
-        s"-Xplugin=${composeProcessor().path}"
-      )
-    } else Seq.empty
-  }
-
-  /**
-   * Enable Jetpack Compose support in the module. Default is `false`.
-   */
-  def androidEnableCompose: T[Boolean] = false
-
-  private def composeProcessor = Task {
-    // cut-off usages for Kotlin 1.x, because of the need to maintain the table of
-    // Compose compiler version -> Kotlin version
-    if (kotlinVersion().startsWith("1"))
-      throw new IllegalStateException("Compose can be used only with Kotlin version 2 or newer.")
-    defaultResolver().resolveDeps(
-      Seq(
-        ivy"org.jetbrains.kotlin:kotlin-compose-compiler-plugin:${kotlinVersion()}"
-      )
-    ).head
-  }
+  
 
   trait AndroidAppKotlinTests extends AndroidAppTests with KotlinTests {
     override def sources: T[Seq[PathRef]] =
